# api/openapi.yaml
openapi: 3.0.3
info:
  title: tfstackplan control-plane API
  version: "1"
  description: |
    The `/api` surface of a tfstackplan tier serve: the runner→server execution
    lifecycle (init → phase → per-stack updates → finalize, plus log ingest and
    the apply-time gate exchange), the apply-lock claim verbs, and the
    execution read.

    This document is the wire contract. It mirrors the shapes in
    `internal/events` (the schemas are bound to those Go types via `x-go-type`
    — there is exactly one canonical definition) and is consumed by code
    generation on both sides: the serve router/stubs and the Go client. Serve
    and runner deploy independently, so changes here are cross-version protocol
    changes: additive only, never a field rename or removal. The byte-level
    truth is pinned by `internal/server/testdata/wire/`.

    Deliberately out of scope: the SSE streams (`GET
    /api/execution/{id}/events`, `/logs/{exec}/{stack}?follow=1` — OpenAPI does
    not model event streams usefully; they are hand-rolled), the GitHub webhook
    and Pub/Sub push endpoints (external contracts), and the public `/img`,
    `/assets`, `/live` surfaces.

    Authentication: every operation requires a Google-signed OIDC ID token
    (bearer). The verified email maps to scopes (`report`, `read`, `admin`) in
    the serve config; each operation lists its accepted scopes (any-of) in
    `x-scopes`. Rejections are `401 unauthorized` / `403 forbidden` as
    `text/plain`.

...
paths:
...
  /api/executions:
    get:
      operationId: listExecutions
      summary: List executions, newest first
      description: |
        The recent-executions feed (default 100, cap 1000). `pr` filters to one
        pull request's full timeline across plan/apply/verify runs. Summaries
        deliberately exclude the rendered report — fetch one execution for
        that.
      security:
        - oidc: [report, read, admin]
      parameters:
        - name: pr
          in: query
          required: false
          schema:
            type: integer
          description: Only executions for this pull request.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
      responses:
        "200":
          description: Execution summaries, newest first. Always an array, never null.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExecutionSummary"
        "400":
          $ref: "#/components/responses/BadRequestText"
        "500":
          $ref: "#/components/responses/InternalText"
...
  /api/execution/{id}:
    get:
      operationId: getExecution
      summary: Read one execution's full state
      description: |
        The execution row, its stack/edge graph, and the gate targets for its
        (pr, environment). NOTE the response's top-level and `gates` keys are
        PascalCase and `CheckRunID` is a raw `{Int64, Valid}` pair — a
        historical accident preserved for wire compatibility.
      security:
        - oidc: [report, read, admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The execution.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionDetail"
        "404":
          description: Unknown execution id.
          content:
            text/plain: {}
        "500":
          $ref: "#/components/responses/InternalText"
components:
  securitySchemes:
    oidc:
      type: http
      scheme: bearer
      bearerFormat: "Google-signed OIDC ID token"
  responses:
    BadRequestText:
      description: Malformed payload (decode/validation error as plain text).
      content:
        text/plain: {}
    InternalText:
      description: Server-side failure (plain-text message).
      content:
        text/plain: {}
  schemas:
    Status:
      type: string
      description: >-
        Per-stack lifecycle status. Decoding rejects unknown non-empty values
        (WIRE-001); an omitted/empty status is tolerated.
      enum:
        - pending
        - initializing
        - initialized
        - running
        - planned
        - gated
        - safe
        - nochange
        - moving
        - failed
        - aborted
      x-go-type: events.Status
      x-go-type-import:
        path: github.com/Fluent-Health/terraform-stack-plan/internal/events
...
    Category:
      type: object
      description: A classification label matched by a stack.
      required: [name]
      properties:
        name:
          type: string
        icon:
          type: string
        attributes:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
      x-go-type: events.Category
      x-go-type-import:
        path: github.com/Fluent-Health/terraform-stack-plan/internal/events
    Counts:
      type: object
      description: A stack's per-kind operation tally (zero counts omitted).
      properties:
        add: { type: integer }
        change: { type: integer }
        destroy: { type: integer }
        replace: { type: integer }
        move: { type: integer }
        import: { type: integer }
        forget: { type: integer }
        sensitivity_only: { type: integer }
      x-go-type: events.Counts
      x-go-type-import:
        path: github.com/Fluent-Health/terraform-stack-plan/internal/events
    StackState:
      type: object
      description: One node in the execution graph.
      required: [path]
      properties:
        path:
          type: string
        project:
          type: string
          description: Grouping/target key (e.g. cloud project/account).
        status:
          $ref: "#/components/schemas/Status"
        detail:
          type: string
          description: Optional failure detail (last error lines).
        categories:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        counts:
          $ref: "#/components/schemas/Counts"
      x-go-type: events.StackState
      x-go-type-import:
        path: github.com/Fluent-Health/terraform-stack-plan/internal/events
    Edge:
      type: object
      description: "A dependency: from must run before to."
      required: [from, to]
      properties:
        from: { type: string }
        to: { type: string }
      x-go-type: events.Edge
      x-go-type-import:
        path: github.com/Fluent-Health/terraform-stack-plan/internal/events
    Graph:
      type: object
      description: The changed subgraph for one execution.
      required: [stacks, edges]
      properties:
        stacks:
          type: array
          items:
            $ref: "#/components/schemas/StackState"
        edges:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/Edge"
      x-go-type: events.Graph
      x-go-type-import:
        path: github.com/Fluent-Health/terraform-stack-plan/internal/events
...
    ExecutionSummary:
      type: object
      description: >-
        One execution in a list view — everything but the graph and the
        rendered report.
      required:
        - id
        - repo
        - sha
        - pr
        - environment
        - context
        - status
        - phase
        - created_at
        - superseded_by
        - log_url
      properties:
        id: { type: string }
        repo: { type: string }
        sha: { type: string }
        pr: { type: integer }
        environment: { type: string }
        context:
          type: string
          description: The commit-status context the run drives (e.g. plan/<env>, apply/<env>, verify/<env>).
        status:
          type: string
          description: Execution-level status (in_progress/success/failure).
        phase:
          type: string
          description: Current lifecycle phase ("" before the first phase event).
        created_at:
          type: string
          format: date-time
        superseded_by:
          type: string
          description: Id of the execution that superseded this one ("" when live).
        log_url: { type: string }
...
    NullInt64:
      type: object
      description: >-
        A database nullable int64, serialized raw (historical accident,
        preserved for wire compatibility).
      required: [Int64, Valid]
      properties:
        Int64:
          type: integer
          format: int64
        Valid:
          type: boolean
    StoredGateTarget:
      type: object
      description: A recorded gate target with its grant state (PascalCase keys).
      required: [Class, Target, GrantName, State, Requester]
      properties:
        Class: { type: string }
        Target: { type: string }
        GrantName:
          type: string
          description: The approval-backend grant resource name (may be empty).
        State:
          type: string
          description: Backend grant state (e.g. REQUESTED, ACTIVE, DENIED, REVOKED).
        Requester:
          type: string
          description: Leased requester SA for this grant.
    ExecutionDetail:
      type: object
      description: >-
        The full execution read. Top-level and gate keys are PascalCase with a
        raw NullInt64 (see getExecution).
      required:
        - ID
        - Repo
        - SHA
        - PR
        - Environment
        - CheckRunID
        - Rev
        - ReportMarkdown
        - LogURL
        - Status
        - StatusContext
        - Phase
        - CreatedAt
        - SupersededBy
        - verify_execution_id
        - graph
        - gates
      properties:
        ID: { type: string }
        Repo: { type: string }
        SHA: { type: string }
        PR: { type: integer }
        Environment: { type: string }
        CheckRunID:
          $ref: "#/components/schemas/NullInt64"
        Rev: { type: integer }
        ReportMarkdown: { type: string }
        LogURL: { type: string }
        Status:
          type: string
          description: Execution-level status (in_progress/success/failure).
        StatusContext: { type: string }
        Phase: { type: string }
        CreatedAt:
          type: string
          format: date-time
        SupersededBy: { type: string }
        verify_execution_id:
          type: string
          description: >-
            Latest verify execution for the same (pr, environment), "" when
            none. Additive (2026-07) — snake_case unlike its legacy siblings.
        graph:
          $ref: "#/components/schemas/Graph"
        gates:
          type: array
          items:
            $ref: "#/components/schemas/StoredGateTarget"